// src/infrastructure/types/generated/MasterFileMap.ts
export interface MasterFileMap {
  /**
   *
   */
  AnimalCodexExcelConfigData: AnimalCodexExcelConfigDataType
  /**
   *
   */
  AvatarCostumeExcelConfigData: AvatarCostumeExcelConfigDataType
  /**
   *
   */
  AvatarCurveExcelConfigData: AvatarCurveExcelConfigDataType
  /**
   *
   */
  AvatarExcelConfigData: AvatarExcelConfigDataType
  /**
   *
   */
  AvatarPromoteExcelConfigData: AvatarPromoteExcelConfigDataType
  /**
   *
   */
  AvatarSkillDepotExcelConfigData: AvatarSkillDepotExcelConfigDataType
  /**
   *
   */
  AvatarSkillExcelConfigData: AvatarSkillExcelConfigDataType
  /**
   *
   */
  AvatarTalentExcelConfigData: AvatarTalentExcelConfigDataType
  /**
   *
   */
  DungeonEntryExcelConfigData: DungeonEntryExcelConfigDataType
  /**
   *
   */
  DungeonLevelEntityConfigData: DungeonLevelEntityConfigDataType
  /**
   *
   */
  EquipAffixExcelConfigData: EquipAffixExcelConfigDataType
  /**
   *
   */
  FetterInfoExcelConfigData: FetterInfoExcelConfigDataType
  /**
   *
   */
  FettersExcelConfigData: FettersExcelConfigDataType
  /**
   *
   */
  FetterStoryExcelConfigData: FetterStoryExcelConfigDataType
  /**
   *
   */
  ManualTextMapConfigData: ManualTextMapConfigDataType
  /**
   *
   */
  MaterialExcelConfigData: MaterialExcelConfigDataType
  /**
   *
   */
  MonsterCurveExcelConfigData: MonsterCurveExcelConfigDataType
  /**
   *
   */
  MonsterDescribeExcelConfigData: MonsterDescribeExcelConfigDataType
  /**
   *
   */
  MonsterExcelConfigData: MonsterExcelConfigDataType
  /**
   *
   */
  ProfilePictureExcelConfigData: ProfilePictureExcelConfigDataType
  /**
   *
   */
  ProudSkillExcelConfigData: ProudSkillExcelConfigDataType
  /**
   *
   */
  ReliquaryAffixExcelConfigData: ReliquaryAffixExcelConfigDataType
  /**
   *
   */
  ReliquaryExcelConfigData: ReliquaryExcelConfigDataType
  /**
   *
   */
  ReliquaryLevelExcelConfigData: ReliquaryLevelExcelConfigDataType
  /**
   *
   */
  ReliquaryMainPropExcelConfigData: ReliquaryMainPropExcelConfigDataType
  /**
   *
   */
  ReliquarySetExcelConfigData: ReliquarySetExcelConfigDataType
  /**
   *
   */
  TowerFloorExcelConfigData: TowerFloorExcelConfigDataType
  /**
   *
   */
  TowerLevelExcelConfigData: TowerLevelExcelConfigDataType
  /**
   *
   */
  TowerScheduleExcelConfigData: TowerScheduleExcelConfigDataType
  /**
   *
   */
  WeaponCurveExcelConfigData: WeaponCurveExcelConfigDataType
  /**
   *
   */
  WeaponExcelConfigData: WeaponExcelConfigDataType
  /**
   *
   */
  WeaponPromoteExcelConfigData: WeaponPromoteExcelConfigDataType
}

/**
 * Type helper to extract decoded type from ExcelBinOutput key
 * Returns array type since all decoded data is an array
 */
export type DecodedType<T extends keyof MasterFileMap> = MasterFileMap[T][]